requirements.txt
APScheduler>=3.8.0
itchat-uos==1.4.1
requests>=2.26.0
